rename stocks.py to app.py + definitions
update gitignore for new structure (defaults)
