+++ stonks/__init__.py
@@ -1,0 +1,4 @@
+from scraper import App, Settings                               # noqa
+from scraper.components.fetchers.base_fetcher import Fetcher    # noqa
+from scraper.components.parsers.base_parser import Parser       # noqa
+from scraper.components.writers.base_writer import Writer       # noqa
