+++ stonks/definitions.py
@@ -1,0 +1,3 @@
+import os
+# Project root
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
